move extend into arrayvec_impl
