+++ src/arrayvec.rs
@@ -707,6 +707,9 @@ impl<T, const CAP: usize> ArrayVecImpl for ArrayVec<T, CAP> {
     const CAPACITY: usize = CAP;
 
     fn len(&self) -> usize { self.len() }
+    fn len_mut(&mut self) -> &mut LenUint {
+        &mut self.len
+    }
 
     unsafe fn set_len(&mut self, length: usize) {
         debug_assert!(length <= CAP);
@@ -987,24 +990,6 @@ impl<'a, T: 'a, const CAP: usize> Drop for Drain<'a, T, CAP> {
     }
 }
 
-struct ScopeExitGuard<T, Data, F>
-    where F: FnMut(&Data, &mut T)
-{
-    value: T,
-    data: Data,
-    f: F,
-}
-
-impl<T, Data, F> Drop for ScopeExitGuard<T, Data, F>
-    where F: FnMut(&Data, &mut T)
-{
-    fn drop(&mut self) {
-        (self.f)(&self.data, &mut self.value)
-    }
-}
-
-
-
 /// Extend the `ArrayVec` with an iterator.
 /// 
 /// ***Panics*** if extending the vector exceeds its capacity.
@@ -1016,75 +1001,6 @@ impl<T, const CAP: usize> Extend<T> for ArrayVec<T, CAP> {
         unsafe {
             self.extend_from_iter::<_, true>(iter)
         }
-    }
-}
-
-#[inline(never)]
-#[cold]
-fn extend_panic() {
-    panic!("ArrayVec: capacity exceeded in extend/from_iter");
-}
-
-impl<T, const CAP: usize> ArrayVec<T, CAP> {
-    /// Extend the arrayvec from the iterable.
-    ///
-    /// ## Safety
-    ///
-    /// Unsafe because if CHECK is false, the length of the input is not checked.
-    /// The caller must ensure the length of the input fits in the capacity.
-    pub(crate) unsafe fn extend_from_iter<I, const CHECK: bool>(&mut self, iterable: I)
-        where I: IntoIterator<Item = T>
-    {
-        let take = self.capacity() - self.len();
-        let len = self.len();
-        let mut ptr = raw_ptr_add(self.as_mut_ptr(), len);
-        let end_ptr = raw_ptr_add(ptr, take);
-        // Keep the length in a separate variable, write it back on scope
-        // exit. To help the compiler with alias analysis and stuff.
-        // We update the length to handle panic in the iteration of the
-        // user's iterator, without dropping any elements on the floor.
-        let mut guard = ScopeExitGuard {
-            value: &mut self.len,
-            data: len,
-            f: move |&len, self_len| {
-                **self_len = len as LenUint;
-            }
-        };
-        let mut iter = iterable.into_iter();
-        loop {
-            if let Some(elt) = iter.next() {
-                if ptr == end_ptr && CHECK { extend_panic(); }
-                debug_assert_ne!(ptr, end_ptr);
-                ptr.write(elt);
-                ptr = raw_ptr_add(ptr, 1);
-                guard.data += 1;
-            } else {
-                return; // success
-            }
-        }
-    }
-
-    /// Extend the ArrayVec with clones of elements from the slice;
-    /// the length of the slice must be <= the remaining capacity in the arrayvec.
-    pub(crate) fn extend_from_slice(&mut self, slice: &[T])
-        where T: Clone
-    {
-        let take = self.capacity() - self.len();
-        debug_assert!(slice.len() <= take);
-        unsafe {
-            let slice = if take < slice.len() { &slice[..take] } else { slice };
-            self.extend_from_iter::<_, false>(slice.iter().cloned());
-        }
-    }
-}
-
-/// Rawptr add but uses arithmetic distance for ZST
-unsafe fn raw_ptr_add<T>(ptr: *mut T, offset: usize) -> *mut T {
-    if mem::size_of::<T>() == 0 {
-        // Special case for ZST
-        (ptr as usize).wrapping_add(offset) as _
-    } else {
-        ptr.add(offset)
     }
 }
