+++ src/arrayvec_impl.rs
@@ -1,7 +1,7 @@
 use std::ptr;
 use std::slice;
 
-use crate::CapacityError;
+use crate::{CapacityError, LenUint};
 
 /// Implements basic arrayvec methods - based on a few required methods
 /// for length and element access.
@@ -82,5 +82,90 @@ pub(crate) trait ArrayVecImpl {
             }
         }
     }
+
+    fn len_mut(&mut self) -> &mut LenUint;
+
+    /// Extend the ArrayVec from the iterable.
+    ///
+    /// ## Safety
+    ///
+    /// Unsafe because if CHECK is false, the length of the input is not checked.
+    /// The caller must ensure the length of the input fits in the capacity.
+    unsafe fn extend_from_iter<I, const CHECK: bool>(&mut self, iterable: I)
+        where I: IntoIterator<Item = Self::Item>
+    {
+        let take = Self::CAPACITY - self.len();
+        let len = self.len();
+        let mut ptr = raw_ptr_add(self.as_mut_ptr(), len);
+        let end_ptr = raw_ptr_add(ptr, take);
+        // Keep the length in a separate variable, write it back on scope
+        // exit. To help the compiler with alias analysis and stuff.
+        // We update the length to handle panic in the iteration of the
+        // user's iterator.
+        let mut guard = ScopeExitGuard {
+            value: self.len_mut(),
+            data: len,
+            f: move |&len, self_len| {
+                **self_len = len as LenUint;
+            }
+        };
+        let mut iter = iterable.into_iter();
+        loop {
+            if let Some(elt) = iter.next() {
+                if ptr == end_ptr && CHECK { extend_panic(); }
+                debug_assert_ne!(ptr, end_ptr);
+                ptr.write(elt);
+                ptr = raw_ptr_add(ptr, 1);
+                guard.data += 1;
+            } else {
+                return; // success
+            }
+        }
+    }
+
+    /// Extend the ArrayVec with copies of elements from the slice;
+    /// the length of the slice must be <= the remaining capacity in the ArrayVec.
+    fn extend_from_slice(&mut self, slice: &[Self::Item])
+    where
+        Self::Item: Clone
+    {
+        let take = Self::CAPACITY - self.len();
+        debug_assert!(slice.len() <= take);
+        unsafe {
+            let slice = if take < slice.len() { &slice[..take] } else { slice };
+            self.extend_from_iter::<_, false>(slice.iter().cloned());
+        }
+    }
 }
 
+#[inline(never)]
+#[cold]
+fn extend_panic() {
+    panic!("ArrayVec: capacity exceeded in extend/from_iter");
+}
+
+struct ScopeExitGuard<T, Data, F>
+    where F: FnMut(&Data, &mut T)
+{
+    pub(crate) value: T,
+    pub(crate) data: Data,
+    pub(crate) f: F,
+}
+
+impl<T, Data, F> Drop for ScopeExitGuard<T, Data, F>
+    where F: FnMut(&Data, &mut T)
+{
+    fn drop(&mut self) {
+        (self.f)(&self.data, &mut self.value)
+    }
+}
+
+/// Rawptr add but uses arithmetic distance for ZST
+unsafe fn raw_ptr_add<T>(ptr: *mut T, offset: usize) -> *mut T {
+    if std::mem::size_of::<T>() == 0 {
+        // Special case for ZST
+        (ptr as usize).wrapping_add(offset) as _
+    } else {
+        ptr.add(offset)
+    }
+}
